ecommerce/requirements.txt
asgiref==3.4.1
Django==3.2.11
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2021.3
sqlparse==0.4.2
